requirements.txt
asgiref==3.7.2
black==23.9.1
bleach==6.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.2
django-environ==0.11.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
webencodings==0.5.1
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1